usando tokens

+++ AULA/express-mongo/services/product.service.mongo.js
@@ -1,8 +1,11 @@
 const ProductModel = require('../models/product.model');
+const auth = require('./auth.service');
+
 
 class ProductService{
 
     static register(req,res){
+        if(!auth.check(req.headers.authorization,res)) return;
         ProductModel.create(req.body).then(
             (prd)=>{
                 res.status(201).json(prd);
@@ -15,6 +18,7 @@ class ProductService{
     }
 
     static list(req,res){
+        if(!auth.check(req.headers.authorization,res)) return;
         ProductModel.find().then(
             (prds)=>{
                 res.status(201).json(prds);
@@ -27,6 +31,7 @@ class ProductService{
     }
 
     static update(req,res){
+        if(!auth.check(req.headers.authorization,res)) return;
         ProductModel.findByIdAndUpdate(req.params.id, req.body, {'new':true}).then(
             (prd)=>{
                 res.status(201).json(prd);
@@ -39,6 +44,7 @@ class ProductService{
     }
 
     static delete(req,res){
+        if(!auth.check(req.headers.authorization,res)) return;
         ProductModel.findByIdAndRemove(req.params.id).then(
             (prd)=>{
                 res.status(201).json(prd);
@@ -51,6 +57,7 @@ class ProductService{
     }
 
     static retrieve(req,res){
+        if(!auth.check(req.headers.authorization,res)) return;
         ProductModel.findById(req.params.id).then(
             (prd)=>{
                 res.status(201).json(prd);
