+++ AULA/express-mongo/services/user.service.mongo.js
@@ -1,24 +1,11 @@
 const UserModel = require('../models/user.model');
-var bcrypt = require('bcrypt');
-var jwt = require('jsonwebtoken');
+const auth = require('./auth.service');
+
 
 class UserService{
 
-    static register(req,res){
-        let rcvUser =  req.body;
-        rcvUser.password = bcrypt.hashSync(rcvUser.password, 10)
-        UserModel.create(rcvUser).then(
-            (user)=>{
-                res.status(201).json(user);
-            }
-        ).catch(
-            (error)=>{
-                res.status(500).json(error);
-            }
-        );
-    }
-
     static list(req,res){
+        if(!auth.check(req.headers.authorization,res)) return;
         UserModel.find().then(
             (users)=>{
                 res.status(201).json(users);
@@ -31,6 +18,7 @@ class UserService{
     }
 
     static update(req,res){
+        if(!auth.check(req.headers.authorization,res)) return;
         UserModel.findByIdAndUpdate(req.params.id, req.body, {'new':true}).then(
             (user)=>{
                 res.status(201).json(user);
@@ -43,6 +31,7 @@ class UserService{
     }
 
     static delete(req,res){
+        if(!auth.check(req.headers.authorization,res)) return;
         UserModel.findByIdAndRemove(req.params.id).then(
             (user)=>{
                 res.status(201).json(user);
@@ -55,6 +44,7 @@ class UserService{
     }
 
     static retrieve(req,res){
+        if(!auth.check(req.headers.authorization,res)) return;
         UserModel.findById(req.params.id).then(
             (user)=>{
                 res.status(201).json(user);
@@ -67,6 +57,7 @@ class UserService{
     }
 
     static retrieveByLogin(req,res){
+        if(!auth.check(req.headers.authorization,res)) return;
         UserModel.find({'login':req.params.login}).then(
             (user)=>{
                 res.status(201).json(user);
@@ -76,39 +67,6 @@ class UserService{
                 res.status(500).json(error);
             }
         );
-    }
-
-    static login(req,res){
-        let loginForm = req.body;
-        UserModel.findOne({'login':loginForm.login})
-        .then(
-            (user)=>{
-                if(bcrypt.compareSync(loginForm.password,user.password)){
-                    //LOGIN ENCONTRADO E CHAVES BATEM
-                    let token = jwt.sign({user: user}, 'secret');
-                    res.status(201).json({
-                        'firstName':user.firstName,
-                        'lastName': user.lastName,
-                        'login':user.login,
-                        'token':token
-                    });
-                }else{
-                    //LOGIN ENCONTRADO MAS CHAVES NÃO BATEM
-                    res.status(201).json(null);
-                }
-            }
-        )
-        .then(undefined,
-            (err)=>{
-                //ERRO EM CASO DE LOGIN NÃO ENCONTRADO
-                res.status(201).json(null);
-            })
-        .catch(
-            (error)=>{
-                res.status(500).json(error);
-            }
-        );
-
     }
 
 }
